+++ lib/menu_item.dart
@@ -2,15 +2,10 @@ import 'package:flutter/material.dart';
 
 import 'base.dart';
 
-enum SMenuItemType { custom, label, switchable }
+enum SMenuItemType { clickable, label, switchable, custom }
 
-/// A Menu Item. Note for the [SBaseDropdownMenu] : It places an InkWell over the
-/// item if it has a value and the onPressed (or onToggle) functions are null
-/// (default). If the value is null or a function is given, the item will no
-/// longer display a preview of the item that is selected and the dropdown
-/// menu will not close when the item is clicked. The onChange function will
-/// not be called. This means that all functionality of the item will have
-/// to be handled by the given function.
+/// A Menu Item. Note for the [SBaseDropdownMenu] : Use a SMenuItem.clickable
+/// to be able to use the onChange function.
 class SMenuItem<T> extends StatelessWidget {
   final SMenuItemType type;
 
@@ -22,21 +17,14 @@ class SMenuItem<T> extends StatelessWidget {
   /// with the showSelected flag set to true.
   final Widget? preview;
 
-  /// Only provide this if you know what you are doing.
-  /// This allows an action to be done by the item.
-  /// Note for the [SBaseDropdownMenu] : It places an InkWell over the
-  /// item if it has a value and the onPressed function is null
-  /// (default). If the value is null or a function is given, the item will no
-  /// longer display a preview of the item that is selected and the dropdown
-  /// menu will not close when the item is clicked. The onChange function will
-  /// not be called. This means that all functionality of the item will have
-  /// to be handled by the given function. It would also be best to use the builder.
+  /// This allows an action to be done by the item. For dropdown menus, this is
+  /// in addition to the onChange function being called.
   final void Function()? onPressed;
 
   // Custom Menu Item properties
   final Widget? child;
-  final Widget Function(BuildContext context, SMenuItemStyle style,
-      Widget? child, void Function()? onPressed)? builder;
+  final Widget Function(
+      BuildContext context, SMenuItemStyle style, Widget? child)? builder;
 
   // Label Menu Item properties
   final Widget? leading;
@@ -46,25 +34,27 @@ class SMenuItem<T> extends StatelessWidget {
   // Switchable Menu Item properties
   final void Function(bool? value)? onToggle;
   final bool? toggled;
-
+  // TODO: add ability to be used in onChange?
   const SMenuItem({
     super.key,
     this.child,
     this.builder,
     this.preview,
-    this.value,
-    this.onPressed,
     this.style = const SMenuItemStyle(),
   })  : leading = null,
         trailing = null,
         title = null,
+        value = null,
+        onPressed = null,
         onToggle = null,
         toggled = null,
-        type = SMenuItemType.custom,
-        assert(!(child == null && builder == null),
-            "SMenuItem error: A child or a builder must be provided.");
+        type = SMenuItemType.custom
+  // TODO: Make this work? SMenuItemButton doesn't work if this is enabled
+  // assert(!(child == null && builder == null),
+  // "SMenuItem error: A child or a builder must be provided."),
+  ;
 
-  const SMenuItem.label({
+  const SMenuItem.clickable({
     super.key,
     this.preview,
     this.value,
@@ -77,28 +67,44 @@ class SMenuItem<T> extends StatelessWidget {
         child = null,
         onToggle = null,
         toggled = null,
+        type = SMenuItemType.clickable;
+
+  const SMenuItem.label({
+    super.key,
+    this.preview,
+    this.leading,
+    this.trailing,
+    this.title,
+    this.style = const SMenuItemStyle(),
+  })  : builder = null,
+        child = null,
+        value = null,
+        onPressed = null,
+        onToggle = null,
+        toggled = null,
         type = SMenuItemType.label;
 
-  // No value since we want to be able to click the toggle
-  // const SMenuItem.switchable({
-  //   super.key,
-  //   required this.onToggle,
-  //   required this.toggled,
-  //   this.style = const SMenuItemStyle(),
-  // })  : leading = null,
-  //       trailing = null,
-  //       title = null,
-  //       preview = null,
-  //       value = null,
-  //       builder = null,
-  //       child = null,
-  //       onPressed = null,
-  //       type = SMenuItemType.switchable;
+  /// WIP
+  const SMenuItem.switchable({
+    super.key,
+    required this.onToggle,
+    required this.toggled,
+    this.title,
+    this.style = const SMenuItemStyle(),
+  })  : leading = null,
+        trailing = null,
+        preview = null,
+        value = null,
+        builder = null,
+        child = null,
+        onPressed = null,
+        type = SMenuItemType.switchable;
 
   // TODO: wrap with Theme to use style.accentColor and style.bgColor
   @override
   Widget build(BuildContext context) {
     switch (type) {
+      case SMenuItemType.clickable:
       case SMenuItemType.label:
         return SizedBox(
           width: style.width,
@@ -108,20 +114,23 @@ class SMenuItem<T> extends StatelessWidget {
             child: Row(
               mainAxisAlignment: style.alignment,
               children: [
-                leading ?? const SizedBox.square(dimension: 10),
+                if (leading != null) leading!,
                 Expanded(child: title ?? Container()),
-                trailing ?? const SizedBox.square(dimension: 10)
+                if (trailing != null) trailing!
               ],
             ),
           ),
         );
       case SMenuItemType.switchable:
         return AnimatedContainer(
-          height: 40,
+          height: style.height,
           margin: const EdgeInsets.only(top: 5),
           duration: const Duration(milliseconds: 250),
+          // TODO: Add checkbox options
           child: CheckboxListTile(
+            activeColor: style.accentColor,
             value: toggled,
+            // TODO: make checkbox work (overlay to new stateful widget?)
             onChanged: onToggle,
             title: title,
           ),
@@ -129,7 +138,7 @@ class SMenuItem<T> extends StatelessWidget {
       case SMenuItemType.custom:
         return builder == null
             ? (child ?? Container())
-            : builder!(context, style, child, onPressed);
+            : builder!(context, style, child);
       default:
         throw Exception(
             "Invalid instantiation of SMenuItem. Use SMenuItem, SMenuItem.label, or SMenuItem.switchable.");